Add the `lint` GitHub Actions job as a reusable workflow
Convert the `lint` job from the GitHub Actions `build.yml` workflow
into a reusable workflow. This will allow all repositories in the org
to use this workflow without needing to define it in each repository.
A number of inputs have been added to ensure the workflow has enough
flexibility since it is no longer defined in a given repository.